requirements.txt
beautifulsoup4==4.13.4
boto3==1.38.3
botocore==1.38.3
numpy==2.2.5
opencv-python-headless==4.11.0.86
pyppeteer==2.0.0
requests==2.32.3
weasyprint==65.1
webvtt-py==0.5.1
python-dotenv==1.1.0
PyYAML==6.0.2